server/service_ai/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "face-verification-service"
version = "1.0.0"
description = "Face verification microservice for attendance system using InsightFace and FAISS"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = {text = "MIT"}
authors = [
    {name = "CIO Verify Face Team", email = "team@cio-verify-face.com"}
]
keywords = [
    "face-recognition",
    "face-verification",
    "attendance",
    "biometric",
    "insightface",
    "arcface",
    "faiss",
    "grpc",
    "fastapi"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    # Core web framework
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    
    # Face recognition and AI
    "insightface==0.7.3",
    "opencv-python==4.8.1.78",
    "numpy==1.26.2",
    "Pillow==10.1.0",
    
    # Vector search
    "faiss-cpu==1.8.0",
    
    # ONNX Runtime (CPU version - default)
    "onnxruntime==1.17.0",
    
    # Database
    "asyncpg==0.29.0",
    "sqlalchemy==2.0.23",
    "alembic==1.12.1",
    
    # HTTP client
    "httpx==0.25.1",
    "aiohttp==3.9.1",
    
    # Monitoring and logging
    "prometheus-client==0.19.0",
    "python-json-logger==2.0.7",
    
    # Security
    "cryptography==41.0.7",
    "python-multipart==0.0.6",
    
    # gRPC
    "grpcio==1.60.0",
    "grpcio-tools==1.60.0",
    "grpcio-reflection==1.60.0",
    
    # Utilities
    "python-dotenv==1.0.0",
    "PyYAML==6.0.1",
]

[project.optional-dependencies]
# GPU support for production
gpu = [
    "onnxruntime-gpu==1.17.0",
]

# Development dependencies
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "black==23.11.0",
    "flake8==6.1.0",
    "mypy==1.7.1",
    "isort==5.12.0",
]

# Testing dependencies
test = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "httpx==0.25.1",
]

# All dependencies
all = [
    "face-verification-service[gpu,dev,test]",
]

[project.urls]
Homepage = "https://github.com/Bot-SomeOne/cio_verify_face"
Documentation = "https://github.com/Bot-SomeOne/cio_verify_face/tree/main/server/service_ai"
Repository = "https://github.com/Bot-SomeOne/cio_verify_face"
Issues = "https://github.com/Bot-SomeOne/cio_verify_face/issues"

[project.scripts]
face-verify = "app.main:main"

[tool.setuptools]
packages = ["app", "app.api", "app.core", "app.models", "app.services", "app.utils"]

[tool.setuptools.package-data]
app = ["*.yaml", "*.yml"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
  | app/grpc_generated
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*", "*/grpc_generated/*"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "insightface.*",
    "faiss.*",
    "cv2.*",
    "grpc.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=app --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/grpc_generated/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "E266", "E501", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
    "*.egg-info",
    ".venv",
    "venv",
    "migrations",
    "grpc_generated",
]
max-complexity = 10
